Mapping out more UI stuff

+++ rangers-api/Hedgehog/System/hhAllocator.h
@@ -19,7 +19,7 @@ namespace hh::fnd
 		inline static void* ms_addrAlloc = sigScan("\x48\x8B\x49\x10\x48\x8B\x01\x45\x85\xC0", "xxxxxxxxxx", (void*)0x140D43FE0);
 		inline static FUNCTION_PTR(void*, __fastcall, ms_fpAlloc, ms_addrAlloc, ThreadSafeTlsfHeapAllocator*, size_t, int);
 
-		csl::fnd::TlsfHeapTemplate<csl::fnd::Mutex> heap;
+        csl::fnd::TlsfHeapTemplate<csl::fnd::Mutex> heap;
 
 	public:
 		~ThreadSafeTlsfHeapAllocator() override
@@ -27,7 +27,7 @@ namespace hh::fnd
 			ms_fpDestructor(this);
 		}
 		
-		void* Alloc(size_t in_size, int in_alignment) override
+		void* Alloc(size_t in_size, size_t in_alignment) override
 		{
 			return ms_fpAlloc(this, in_size, in_alignment);
 		}
