+++ rangers-api/Sonicteam/System/IAllocator.h
@@ -7,7 +7,7 @@ namespace csl::fnd
 	public:
 		virtual ~IAllocator() = default;
 		
-		virtual void* Alloc(size_t in_size, int in_alignment) = 0;
+		virtual void* Alloc(size_t in_size, size_t in_alignment) = 0;
 		virtual void Free(void* in_pMemory) = 0;
 	};
 }
